// src/common/enums/action.enum.ts
export enum Action {
  CREATE = 'create',
  READ = 'read',
  UPDATE = 'update',
  DELETE = 'delete',
  MANAGE = 'manage',
  //! SET_PERMISSIONS = 'set_permission',
  // READ_OWN = 'read_own',
  // UPDATE_OWN = 'update_own',
  // DELETE_OWN = 'delete_own',
  // MANAGE_OWN = 'manage_own',
  //!SPECIAL ACTION
}